add serde_json_path, create template_renderer.rs
