+++ Cargo.toml
@@ -23,6 +23,7 @@ pulldown-cmark = "0.9.3"
 regex = "1.10.1"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
+serde_json_path = "0.6.4"
 serde_regex = "1.1.0"
 serde_yaml = "0.9.25"
 tracing = "0.1.39"
